Issue #29 - Rework pre-commit handling in preparation for adding tox
Rework the way we're handling checks in pre-commit in preparation for
adding tox to our tooling. We will be calling our pre-commit "linters"
from tox so we need to prep a few things so it cooperates:

- Move as many checks as possible over to pull specific revisions of the
repositories. The caching in `pre-commit` is much nicer and saves us the
time of `tox` building a new environment, etc. Note, these specific revs
should be updated regularly via `pre-commit autoupdate`.

- Add filtering to keep some of checks from running across the
repository.

- Re-baseline our repository with a run of `pre-commit run
--all-files`. This is the same thing `tox` is going to use to call our
steps.