Update core pre-commit linting steps to use local dependencies
Update pre-commit config so black, mypy, and flake8 run `local`
dependencies instead of using their own virtual environment. This
addresses the issue raised in #43 with additional dependencies for checks
being duplicated across `setup.cfg` and `.pre-commit-config.yaml`.

Update `mypy` and `flake8` calls to explicitly check `src` instead of
being passed individual file names to validate. For `mypy` this ensures
that we're doing a complete scan that is identical to a CLI call. For
`flake8` this just keeps our code quality nagging to the core code.

Resolve #43